# openapi.yaml
openapi: 3.0.3
info:
  title: Memory Manager API
  description: API for managing memories and LLM capabilities
  version: 1.0.0
servers:
  - url: http://localhost:8000
    description: Local development server
paths:
...
  /memories:
    get:
      summary: Get all memory abstracts
      description: Retrieve all memory abstracts from the repository
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  memories:
                    type: array
                    items:
                      type: object
                      properties:
                        name:
                          type: string
                        abstract:
                          type: string
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
    post:
      summary: Add a new memory
      description: Add a new memory to the repository
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - abstract
                - memory_block
              properties:
                name:
                  type: string
                abstract:
                  type: string
                memory_block:
                  type: string
      responses:
        '201':
          description: Memory created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
  /memories/{name}:
    get:
      summary: Get a specific memory
      description: Retrieve a specific memory by name
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  name:
                    type: string
                  abstract:
                    type: string
                  memory_block:
                    type: string
        '404':
          description: Memory not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
    put:
      summary: Update an existing memory
      description: Update an existing memory by name
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - abstract
                - memory_block
              properties:
                abstract:
                  type: string
                memory_block:
                  type: string
      responses:
        '200':
          description: Memory updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '404':
          description: Memory not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
    delete:
      summary: Delete a memory
      description: Delete a memory by name
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Memory deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '404':
          description: Memory not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
  /visible-memories:
    get:
      summary: Get visible memories
      description: Retrieve all visible memories in the current manager
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  visible_memories:
                    type: array
                    items:
                      type: object
                      properties:
                        name:
                          type: string
                        abstract:
                          type: string
                        memory_block:
                          type: string
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
...
  /full-update:
    post:
      summary: Perform a full update cycle
      description: Create new memories, update existing ones, and update relevance
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - chat_messages
                - delta
              properties:
                chat_messages:
                  type: array
                  items:
                    type: object
                    properties:
                      role:
                        type: string
                      text:
                        type: string
                delta:
                  type: integer
      responses:
        '200':
          description: Full update completed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  visible_memories:
                    type: array
                    items:
                      type: object
                      properties:
                        name:
                          type: string
                        abstract:
                          type: string
                        memory_block:
                          type: string
                  relevance_map:
                    type: object
                    additionalProperties:
                      type: integer
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
  /update-existing-memories:
    post:
      summary: Update all memories
      description: Update all memories in the manager using LLM capabilities
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - chat_messages
              properties:
                chat_messages:
                  type: array
                  items:
                    type: object
                    properties:
                      role:
                        type: string
                      text:
                        type: string
      responses:
        '200':
          description: All memories updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  visible_memories:
                    type: array
                    items:
                      type: object
                      properties:
                        name:
                          type: string
                        abstract:
                          type: string
                        memory_block:
                          type: string
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
  /extract-new-memories:
    post:
      summary: Create new memories
      description: Create new memories based on chat messages and existing memories
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - chat_messages
              properties:
                chat_messages:
                  type: array
                  items:
                    type: object
                    properties:
                      role:
                        type: string
                      text:
                        type: string
      responses:
        '200':
          description: New memories created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  memories:
                    type: array
                    items:
                      type: object
                      properties:
                        name:
                          type: string
                        abstract:
                          type: string
                        memory_block:
                          type: string
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
  /refresh-visible-memory-list:
    post:
      summary: Refresh visible memories
      description: Refresh visible memories based on relevance counts
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - limit
              properties:
                limit:
                  type: integer
      responses:
        '200':
          description: Visible memories list refreshed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  visible_memories:
                    type: array
                    items:
                      type: object
                      properties:
                        name:
                          type: string
                        abstract:
                          type: string
                        memory_block:
                          type: string
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
  /update-relevance-map:
    post:
      summary: Update relevance map
      description: Update relevance map based on memories related to chat messages
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - chat_messages
                - delta
              properties:
                chat_messages:
                  type: array
                  items:
                    type: object
                    properties:
                      role:
                        type: string
                      text:
                        type: string
                delta:
                  type: integer
      responses:
        '200':
          description: Relevance map updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  relevance_map:
                    type: object
                    additionalProperties:
                      type: integer
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
  /update-visible-memory-list:
    post:
      summary: Update visible memories based on relevance
      description: Update visible memories based on relevance to chat messages
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - chat_messages
                - limit
              properties:
                chat_messages:
                  type: array
                  items:
                    type: object
                    properties:
                      role:
                        type: string
                      text:
                        type: string
                limit:
                  type: integer
                delta:
                  type: integer
                  default: 1
      responses:
        '200':
          description: Visible memories updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  relevance_map:
                    type: object
                    additionalProperties:
                      type: integer
                  visible_memories:
                    type: array
                    items:
                      type: object
                      properties:
                        name:
                          type: string
                        abstract:
                          type: string
                        memory_block:
                          type: string
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
  /force-update-relevance-map:
    post:
      summary: Force update relevance map
      description: Update the relevance map with delta values
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - delta_map
              properties:
                delta_map:
                  type: object
                  additionalProperties:
                    type: integer
      responses:
        '200':
          description: Relevance map updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  relevance_map:
                    type: object
                    additionalProperties:
                      type: integer
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
...